resolved TypeScript issues
