+++ site/src/layouts/DetailComponent/index.ts
@@ -1,2 +1,1 @@
 export * from "./DetailComponent";
-export * from "./useAllExamplesView";
