--- site/src/layouts/DetailComponent/useAllExamplesView.ts
@@ -1,9 +1,0 @@
-import { useContext } from "react";
-import {
-  AllExamplesViewContextType,
-  AllExamplesViewContext,
-} from "./DetailComponent";
-
-export const useAllExamplesView = (): AllExamplesViewContextType => {
-  return useContext(AllExamplesViewContext);
-};
